Client: Moving around and renaming Create to Movie. Next planned step is to add edit movie ability, so having it named "Create" doesn't make too much sense, does it?
