+++ client/src/components/movie/Basics.js
